Refactored code and introduced bugs
Refactored code for Smithy, Adventurer, Mine, Feast and Council Room so that they are implemented in their own functions rather than as part of the switch statement in cardEffect. The functions are called based on the switch conditionals. Four bugs were also introduced, one in each of these except for Mine.